move function GetKubemarkMasterComponentResoureUsage
